+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     * Updated schema and XSLT for ui:abbr to change attribute `description` to `toolTip` in line with other components.
 * Method scope issue in WCardManager. Removed "removeAll(flag)" and "remove(child, flag)". (#59)
 * "WebUtilities.updateBeanValue(component)" changed to ignore invisible components by default. "WebUtilities.updateBeanValue(component, flag)" added to provide this option. (#46)
+* Removed deprecated com.github.bordertech.wcomponents.test.util.TreeUtil. Use com.github.bordertech.wcomponents.util.TreeUtil instead.
+* Removed deprecated com.github.bordertech.wcomponents.test.util.WComponentTreeVisitor. Use com.github.bordertech.wcomponents.util.WComponentTreeVisitor instead.
 
 ## Enhancements
 * WTabSet implements SubordinateTarget. (#159)